requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
